* Removed SetWheelAngularVelocity from HAL. HAL-level wheel driving function is MotorSetPower.    Replaced with call to new SteeringController::ExecuteDirectDrive(leftVel, rightVel).
* Added a SteerMode to SteeringController which determines how the wheels are being driven. Might eventually want to collapse these controllers together or maybe consolidate this with Robot::OperationMode(?)

  i) Path following
     Controller tries to reduce distance error and angular error to path obtained from PathFollower. Speed determined by SetUserCommandedDesiredSpeed().

  ii) Point turning:
      Triggered by ExecutePointTurn. A special case of path following. In a pony turn, speed is technically 0 so speeds are specified as parameters to this command overriding userCommandedDesiredSpeed.

  iii) Direct drive:
      Control individual wheels speeds explicitly with ExecuteDirectDrive. Speeds are specified as parameters to this command overriding userCommandedDesiredSpeed.

* Removed Speed, Steer, Wheel Controller Manage() calls out of PathFollower since they are not specific to path following.

* PathFollower
  - Changed float to f32
  - Added point turn logic
  - Update() now does most of what GetPathError() used to do. GetPathError() now only returns errors, as its name implies.
    SteerController now gets the errors from PathFollower when it's in PathFollower mode.
  - Fixed some bugs

* WheelController
  - Change desired wheel speeds from s16 to f32

* DockingController
  - Replaced call to SetWheelAngularVelocity with ExecuteDirectDrive. This almost certainly breaks docking.

+++ include/anki/cozmo/robot/wheelController.h
@@ -51,8 +51,8 @@ namespace Anki {
     void Manage();
     
     //Sets/Gets the desired speeds for the wheels (in mm/sec forward speed)
-    void GetDesiredWheelSpeeds(s16 *leftws, s16 *rightws);
-    void SetDesiredWheelSpeeds(s16 leftws, s16 rightws);
+    void GetDesiredWheelSpeeds(f32 *leftws, f32 *rightws);
+    void SetDesiredWheelSpeeds(f32 leftws, f32 rightws);
     
     void GetFilteredWheelSpeeds(f32 *left, f32 *right);
